+++ daml-assistant/src/DAML/Assistant/Types.hs
@@ -44,6 +44,7 @@ assistantErrorBecause msg e = (assistantError msg) { errInternal = Just e }
 data Env = Env
     { envDamlPath      :: DamlPath
     , envDamlAssistantPath :: DamlAssistantPath
+    , envDamlAssistantSdkVersion :: Maybe DamlAssistantSdkVersion
     , envProjectPath   :: Maybe ProjectPath
     , envSdkPath       :: Maybe SdkPath
     , envSdkVersion    :: Maybe SdkVersion