+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/sandbox/config/SandboxConfig.scala
@@ -20,7 +20,7 @@ final case class TlsServerConfiguration(
   * Defines the basic configuration for running sandbox
   */
 final case class SandboxConfig(
-    addressOption: Option[String],
+    address: Option[String],
     port: Int,
     damlPackageContainer: DamlPackageContainer,
     timeProviderType: TimeProviderType,