+++ navigator/backend/src/main/scala/com/digitalasset/navigator/console/commands/Sql.scala
@@ -23,8 +23,9 @@ case object Sql extends SimpleCommand {
       ps <- state.getPartyState ~> s"Unknown party ${state.party}"
       result <- ps.ledger.runQuery(query) ~> s"Error while running the query"
     } yield {
+      val width = state.reader.getTerminal.getWidth
       val table = AsciiTable()
-        .width(state.reader.getTerminal.getWidth)
+        .width(if (width > 4) width else 80)
         .multiline(true)
         .columnMinWidth(4)
         .sampleAtMostRows(100000)